requirements.txt
Flask==2.1.3
sling
PyYAML==6.0
Werkzeug==2.0.3
flask-cors
simple-salesforce
pandas
boto3
pymongo